代码规范整理

+++ Admin/PartnerAdminGroupDestroyAction.php
@@ -46,18 +46,20 @@ class PartnerAdminGroupDestroyAction
             $datas->delete();
             //检查是否有人工充值权限
             $role = $datas->role == '*' ? Arr::wrap($datas->role) : Arr::wrap(json_decode($datas->role, true));
-            $fundOperation = BackendSystemMenu::select('id')->where('route', '/manage/recharge')->first()->toArray();
-            $isManualRecharge = in_array($fundOperation['id'], $role, true);
-            //如果有有人工充值权限   删除  FundOperation  BackendAdminRechargePermitGroup 表
-            if ($isManualRecharge === true) {
-                $fundOperationGroup = new BackendAdminRechargePermitGroup();
-                $fundOperationGroup->where('group_id', $id)->delete();
-                //需要删除的资金表 admin
-                $fundOperationEloq = new BackendAdminRechargePocessAmount();
-                $adminsData = BackendAdminUser::select('id')->where('group_id', $id)->get();
-                $admins = array_column($adminsData->toArray(), 'id');
-                if ($adminsData !== null) {
-                    $fundOperationEloq->whereIn('admin_id', $admins)->delete();
+            $fundOperation = BackendSystemMenu::select('id')->where('route', '/manage/recharge')->first();
+            if ($fundOperation !== null) {
+                $isManualRecharge = in_array($fundOperation['id'], $role, true);
+                //如果有有人工充值权限   删除  FundOperation  BackendAdminRechargePermitGroup 表
+                if ($isManualRecharge === true) {
+                    $fundOperationGroup = new BackendAdminRechargePermitGroup();
+                    $fundOperationGroup->where('group_id', $id)->delete();
+                    //需要删除的资金表 admin
+                    $fundOperationEloq = new BackendAdminRechargePocessAmount();
+                    $adminsData = BackendAdminUser::select('id')->where('group_id', $id)->get();
+                    $admins = array_column($adminsData->toArray(), 'id');
+                    if ($adminsData !== null) {
+                        $fundOperationEloq->whereIn('admin_id', $admins)->delete();
+                    }
                 }
             }
             return $contll->msgOut(true);
